Switch to internal networking
Host-only interfaces are no longer supported by Virtualbox. The
host-only network feature described in the green box over at
https://www.virtualbox.org/manual/ch06.html#network_hostonly is not yet
supported by Vagrant as it seems.

With internal networking port forwarding becomes necessary as the host
is no longer in the same network.

With the "Host-Only Network" feature this would be the case again.

Additionally the file has been tidied up.